Endpoints bank of resume

+++ internal/services/bank_of_resume_service.go
@@ -1,22 +1,23 @@
 package services
 
 import (
-	"context"
 	gormmanagers "lamsam-web3-backend/internal/adapter/gorm/managers"
 	"lamsam-web3-backend/internal/customerrors"
 	"lamsam-web3-backend/internal/dto"
 	"lamsam-web3-backend/internal/models"
 	"lamsam-web3-backend/internal/repositories"
+	"lamsam-web3-backend/internal/utils"
 
+	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
 )
 
 type BankOfResumeService interface {
 	CreateBankOfResume(bankOfResume *models.BankOfResume, userID uint, subtopicIDs []uint) (*models.BankOfResume, error)
-	UpdateBankOfResume(bankOfResume *models.BankOfResume) error
-	DeleteBankOfResume(id uint) error
+	UpdateBankOfResume(bankOfResume *models.BankOfResume, userID uint, role string) error
+	DeleteBankOfResume(id uint, userId uint, role string) error
 	GetBankOfResumeByID(id uint) (*models.BankOfResume, error)
-	GetAllBankOfResumes(ctx context.Context) ([]models.BankOfResume, *dto.PaginationDTO, error)
+	GetAllBankOfResumes(c *gin.Context) (*dto.PaginationDTO, error)
 }
 
 type bankOfResumeService struct {
@@ -62,17 +63,42 @@ func (s *bankOfResumeService) CreateBankOfResume(bankOfResume *models.BankOfResu
 	return createdBankOfResume, nil
 }
 
-func (s *bankOfResumeService) UpdateBankOfResume(bankOfResume *models.BankOfResume) error {
+func (s *bankOfResumeService) UpdateBankOfResume(bankOfResume *models.BankOfResume, userID uint, role string) error {
 	if bankOfResume == nil || bankOfResume.ID == 0 {
 		return customerrors.ErrInvalidData
 	}
+
+	existing, err := s.GetBankOfResumeByID(bankOfResume.ID)
+	if err != nil {
+		return err
+	}
+
+	if existing.UserID != userID && role != "admin" {
+		return customerrors.ErrUnauthorized
+	}
+
+	updates := utils.GetModifiedFields(existing, bankOfResume)
+	if len(updates) == 0 {
+		return nil
+	}
+
 	return s.repo.Update(bankOfResume)
 }
 
-func (s *bankOfResumeService) DeleteBankOfResume(id uint) error {
+func (s *bankOfResumeService) DeleteBankOfResume(id uint, userID uint, role string) error {
 	if id == 0 {
 		return customerrors.ErrInvalidID
 	}
+
+	existing, err := s.GetBankOfResumeByID(id)
+	if err != nil {
+		return err
+	}
+
+	if existing.UserID != userID && role != "admin" {
+		return customerrors.ErrUnauthorized
+	}
+
 	return s.repo.Delete(id)
 }
 
@@ -83,6 +109,6 @@ func (s *bankOfResumeService) GetBankOfResumeByID(id uint) (*models.BankOfResume
 	return s.repo.GetByID(id)
 }
 
-func (s *bankOfResumeService) GetAllBankOfResumes(ctx context.Context) ([]models.BankOfResume, *dto.PaginationDTO, error) {
-	return s.repo.GetAll(ctx)
+func (s *bankOfResumeService) GetAllBankOfResumes(c *gin.Context) (*dto.PaginationDTO, error) {
+	return s.repo.GetAll(c)
 }
